fix(ai): emit strict-compatible agent schemas
