+++ packages/common/src/ee/AiAgent/schemas/toolDefinitionWithoutMcpOutput.ts
@@ -1,5 +1,6 @@
 import snakeCase from 'lodash/snakeCase';
 import { type z } from 'zod';
+import { createAgentInputSchema } from './agentInputSchema';
 import {
     type AgentToolConfig,
     type AgentToolView,
@@ -11,7 +12,6 @@ import {
 import { createMcpCompatibleInputSchema } from './McpSchemaCompatLayer';
 import {
     assertAvailable,
-    createAgentInputSchema,
     createMcpToolResultBuilders,
     defaultAgentToModelOutput,
     resolveDescription,
